Refactor charts services
