+++ test/unit/charts/oem_distribution_test.rb
@@ -7,18 +7,19 @@ module ForemanOmaha
         User.current = FactoryBot.create(:user, :admin)
       end
 
-      let(:oem_distribition_chart) { ForemanOmaha::Charts::OemDistribution.new }
+      let(:omaha_group) { FactoryBot.create(:omaha_group) }
+      let(:oem_distribition_chart) { ForemanOmaha::Charts::OemDistribution.new(omaha_group) }
 
       context 'with hosts' do
         setup do
-          FactoryBot.create_list(:host, 5, :with_omaha_facet)
+          FactoryBot.create_list(:host, 5, :with_omaha_facet, omaha_group: omaha_group)
         end
 
         test 'calculates the oem distribution' do
-          expected = [
-            { :label => 'rackspace', :data => 5 }
-          ]
-          assert_equal(expected, oem_distribition_chart.to_chart_data.sort_by { |e| e[:label] })
+          expected = [['rackspace', 5]]
+          actual = JSON.parse(oem_distribition_chart.to_chart_data)['columns']
+
+          assert_equal expected, actual
         end
       end
     end
